pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-cost-optimizer"
version = "0.1.0"
description = "LLM cost management system with smart request routing based on task complexity"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Onurcan Donmezer"},
]
keywords = ["llm", "cost-optimization", "routing", "ai", "budget-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "streamlit>=1.28.0",
    "pandas>=2.0",
    "plotly>=5.15.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/onurcandnmz/llm-cost-optimizer"
Repository = "https://github.com/onurcandnmz/llm-cost-optimizer"
Issues = "https://github.com/onurcandnmz/llm-cost-optimizer/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["src/app.py"]

[tool.coverage.report]
show_missing = true
fail_under = 80
